null

+++ Assets/MiniFramework/04.ResKit/HttpDownloader.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Collections;
 using System.Collections.Generic;
 using System.IO;
 using System.Net;
@@ -13,7 +12,7 @@ namespace MiniFramework
         Stop,
         Completed
     }
-    public class HttpDownloader
+    public class HttpDownloader:Singleton<HttpDownloader>
     {
         private Dictionary<int,DownloadTask> tasks = new Dictionary<int, DownloadTask>();
         public DownState GetStatus(int id)
@@ -24,25 +23,30 @@ namespace MiniFramework
         {
             return tasks[id].GetProcess();
         }
-        public void Download(int id, MonoBehaviour mono, string url, string savePath, Action callBack = null)
+        private HttpDownloader() { }
+        public void Download(int id,string url, string savePath, Action callBack = null)
         {       
             if (!tasks.ContainsKey(id))
             {
-                DownloadTask task = new DownloadTask(mono, url, savePath, callBack);
+                DownloadTask task = new DownloadTask(url, savePath, callBack);
                 tasks.Add(id,task);
             }
-        }     
+        }
     }
     public class DownloadTask
-
     {
         private DownState state;
-        private string url;
-        private string tempSavePath;
-        private string savePath;
-        private long curLength;
-        private long fileLength;
+        private string url;//下载地址
+        private string tempSavePath;//临时文件保存路径
+        private string savePath;//下载文件保存路径
+        private long curLength;//临时文件长度
+        private long fileLength;//下载文件总长度
         private Action CallBack;
+
+        private FileStream fileStream;
+        private HttpWebResponse response;
+        private Stream responseStream;
+        private byte[] buffer = new byte[1024];
         public DownState GetStatus()
         {
             return state;
@@ -55,21 +59,21 @@ namespace MiniFramework
             }
             return 0;
         }
-        public  DownloadTask(MonoBehaviour mono, string url, string savePath, Action callBack = null)
+        public  DownloadTask(string url, string savePath, Action callBack = null)
         {
             state = DownState.None;
             this.url = url;
             tempSavePath = savePath + ".temp";
             this.savePath = savePath;
             CallBack = callBack;
-            mono.StartCoroutine(Task());
+            DownTask();
         }
-        IEnumerator Task()
+
+        void DownTask()
         {
             state = DownState.Start;
             HttpWebRequest request = (HttpWebRequest)HttpWebRequest.Create(url);
             request.Method = "GET";
-            FileStream fileStream;
             if (File.Exists(tempSavePath))
             {
                 fileStream = File.OpenWrite(tempSavePath);
@@ -83,39 +87,44 @@ namespace MiniFramework
                 curLength = 0;
             }
             request.KeepAlive = false;
-            WebResponse response = (HttpWebResponse)request.GetResponse();
-            Stream stream = response.GetResponseStream();
-            fileLength = response.ContentLength + curLength;
-            int lengthOnce;
-            long bufferMaxLength = response.ContentLength;
-            while (curLength < fileLength)
-            {
-                byte[] buffer = new byte[bufferMaxLength];
-                if (stream.CanRead)
-                {
-                    lengthOnce = stream.Read(buffer, 0, buffer.Length);
-                    curLength += lengthOnce;
-                    fileStream.Write(buffer, 0, lengthOnce);
-                }
-                else
-                {
-                    break;
-                }
-                yield return new WaitForFixedUpdate();
-            }
-            response.Close();
-            stream.Close();
-            fileStream.Close();
-            if (File.Exists(savePath))
-            {
-                File.Delete(savePath);
-            }
-            File.Move(tempSavePath, savePath);
-            if (CallBack != null)
-            {
-                CallBack();
-            }
-            state = DownState.Completed;
+            request.BeginGetResponse(new AsyncCallback(RespCallBack), request);
         }
+        void RespCallBack(IAsyncResult result)
+        {
+            HttpWebRequest request = (HttpWebRequest)result.AsyncState;
+            response = (HttpWebResponse)request.EndGetResponse(result);
+            responseStream = response.GetResponseStream();
+            fileLength = response.ContentLength + curLength;
+            responseStream.BeginRead(buffer, 0, buffer.Length,new AsyncCallback(ReadCallBack), responseStream);
+        }
+
+        void ReadCallBack(IAsyncResult result)
+        {
+            responseStream = (Stream)result.AsyncState;
+            int read = responseStream.EndRead(result);
+            curLength += read;
+            if (read > 0)
+            {
+                fileStream.Write(buffer,0,read);
+                responseStream.BeginRead(buffer, 0, buffer.Length, new AsyncCallback(ReadCallBack), responseStream);
+            }
+            else
+            {
+                fileStream.Close();
+                responseStream.Close();
+                response.Close();
+                if (File.Exists(savePath))
+                {
+                    File.Delete(savePath);
+                }
+                File.Move(tempSavePath, savePath);
+                if (CallBack != null)
+                {
+                    CallBack();
+                }
+                state = DownState.Completed;
+            }
+        }
+
     }
 }
